requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.3
duckdb==0.9.1
openai>=1.12.0
python-dotenv==1.0.0
lxml==4.9.3
html5lib==1.1
pillow==10.0.1
urllib3==2.0.4
gunicorn==21.2.0